+++ CHANGELOG.md
@@ -20,6 +20,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   name for each project on the projects index page.
 - Project contacts can be added and edited
 - Add hint text and guidance text to actions.
+- A regional delivery officer must be selected at the time of project creation.
+  A regional delivery officer can only see project they are assigned to on the
+  projects index page.
 
 ### Removed